attempt to fix violation
