--- src/main/java/com/app/folioman/config/Scheme.java
@@ -1,14 +1,0 @@
-package com.app.folioman.config;
-
-public class Scheme {
-
-    private String dataUrl;
-
-    public String getDataUrl() {
-        return dataUrl;
-    }
-
-    public void setDataUrl(String dataUrl) {
-        this.dataUrl = dataUrl;
-    }
-}
