+++ src/main/java/com/app/folioman/mfschemes/config/BseStar.java
@@ -1,0 +1,21 @@
+package com.app.folioman.mfschemes.config;
+
+import jakarta.validation.Valid;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.validation.annotation.Validated;
+
+@Validated
+@ConfigurationProperties(prefix = "app.bsestar")
+public class BseStar {
+
+    @Valid
+    private Scheme scheme;
+
+    public Scheme getScheme() {
+        return scheme;
+    }
+
+    public void setScheme(Scheme scheme) {
+        this.scheme = scheme;
+    }
+}
