+++ src/main/java/com/app/folioman/mfschemes/service/AmfiService.java
@@ -1,6 +1,6 @@
 package com.app.folioman.mfschemes.service;
 
-import com.app.folioman.config.ApplicationProperties;
+import com.app.folioman.mfschemes.config.ApplicationProperties;
 import com.opencsv.CSVReader;
 import com.opencsv.exceptions.CsvException;
 import java.io.IOException;
